Add functionality to digits, operations and clear
The calculator can do the basic operations of addition, subtraction, multiplication and division. It can also clear the display and start a new calculation. The logic of how this is done is through checking the existence or not of the variables of the numbers and the operator. There are only a few possible cases and each of them can be handled separately.