Add decimal button
The decimal button is programmed to work as a digit button with the difference that we need to check if it is the first button press in a new number or not. If it is the first button press, we need to add a 0 before the decimal point. If it is not the first button press, we need to check if there is already a decimal point in the number. If there is, we should not add another one.